Move backports to their own file, to more clearly separate them from utilities
By starting this new module name with a `_`, we indicate that users are not supposed to be touching these pieces, and they will be removed without notice (when cocotb increases its minimum python version.
