+++ cocotb/regression.py
@@ -54,8 +54,9 @@ import cocotb
 import cocotb.ANSI as ANSI
 from cocotb.log import SimLog
 from cocotb.result import TestError, TestFailure, TestSuccess, SimFailure
-from cocotb.utils import get_sim_time, raise_from
+from cocotb.utils import get_sim_time
 from cocotb.xunit_reporter import XUnitReporter
+from cocotb import _py_compat
 
 
 def _my_import(name):
@@ -148,7 +149,7 @@ class RegressionManager(object):
                     except AttributeError:
                         self.log.error("Requested test %s wasn't found in module %s", test, module_name)
                         err = AttributeError("Test %s doesn't exist in %s" % (test, module_name))
-                        raise_from(err, None)  # discard nested traceback
+                        _py_compat.raise_from(err, None)  # discard nested traceback
 
                     if not hasattr(_test, "im_test"):
                         self.log.error("Requested %s from module %s isn't a cocotb.test decorated coroutine", test, module_name)
