+++ cocotb/utils.py
@@ -44,18 +44,13 @@ else:
     simulator = None
     _LOG_SIM_PRECISION = -15
 
-# This is six.integer_types
-if sys.version_info.major >= 3:
-    integer_types = (int,)
-else:
-    integer_types = (int, long)  # noqa
-
 
 def get_python_integer_types():
     warnings.warn(
         "This is an internal cocotb function, use six.integer_types instead",
         DeprecationWarning)
-    return integer_types
+    from cocotb import _py_compat
+    return _py_compat.integer_types
 
 
 # Simulator helper functions
@@ -402,79 +397,6 @@ def hexdiffs(x, y):
     return rs
 
 
-# This is essentially six.exec_
-if sys.version_info.major == 3:
-    # this has to not be a syntax error in py2
-    import builtins
-    exec_ = getattr(builtins, 'exec')
-else:
-    # this has to not be a syntax error in py3
-    def exec_(_code_, _globs_=None, _locs_=None):
-        """Execute code in a namespace."""
-        if _globs_ is None:
-            frame = sys._getframe(1)
-            _globs_ = frame.f_globals
-            if _locs_ is None:
-                _locs_ = frame.f_locals
-            del frame
-        elif _locs_ is None:
-            _locs_ = _globs_
-        exec("""exec _code_ in _globs_, _locs_""")
-
-
-# this is six.with_metaclass, with a clearer docstring
-def with_metaclass(meta, *bases):
-    """This provides:
-
-    .. code-block:: python
-
-        class Foo(with_metaclass(Meta, Base1, Base2)): pass
-
-    which is a unifying syntax for:
-
-    .. code-block:: python
-
-        # Python 3
-        class Foo(Base1, Base2, metaclass=Meta): pass
-
-        # Python 2
-        class Foo(Base1, Base2):
-            __metaclass__ = Meta
-    """
-    # This requires a bit of explanation: the basic idea is to make a dummy
-    # metaclass for one level of class instantiation that replaces itself with
-    # the actual metaclass.
-    class metaclass(type):
-
-        def __new__(cls, name, this_bases, d):
-            return meta(name, bases, d)
-
-        @classmethod
-        def __prepare__(cls, name, this_bases):
-            return meta.__prepare__(name, bases)
-    return type.__new__(metaclass, 'temporary_class', (), {})
-
-
-# this is six.raise_from
-if sys.version_info[:2] == (3, 2):
-    exec_("""def raise_from(value, from_value):
-    try:
-        if from_value is None:
-            raise value
-        raise value from from_value
-    finally:
-        value = None
-    """)
-elif sys.version_info[:2] > (3, 2):
-    exec_("""def raise_from(value, from_value):
-    try:
-        raise value from from_value
-    finally:
-        value = None
-    """)
-else:
-    def raise_from(value, from_value):
-        raise value
 
 
 class ParametrizedSingleton(type):
@@ -509,26 +431,6 @@ class ParametrizedSingleton(type):
             cls.__instances[key] = self
             return self
 
-
-# backport of Python 3.7's contextlib.nullcontext
-class nullcontext(object):
-    """Context manager that does no additional processing.
-    Used as a stand-in for a normal context manager, when a particular
-    block of code is only sometimes used with a normal context manager:
-
-    cm = optional_cm if condition else nullcontext()
-    with cm:
-        # Perform operation, using optional_cm if condition is True
-    """
-
-    def __init__(self, enter_result=None):
-        self.enter_result = enter_result
-
-    def __enter__(self):
-        return self.enter_result
-
-    def __exit__(self, *excinfo):
-        pass
 
 
 def reject_remaining_kwargs(name, kwargs):
